bug(dependencies): Add imports of demos
Demos of elements like table and list import other elements
which are not installed along with the element so we need
to install them and import them in the application. This is
not the best and long term policy we should adopt. Like in the
case of elements browser we should find out the demo imports
and handle them at the site creation time.